:truck: Rename 'deep-merge' files to 'merge-deep'
This is also how function is named. It seemed unintuitive to name it the other way around.
